updated dataset compatibility with original code

+++ training/ddp_training/modules_ddp.py
@@ -566,7 +566,7 @@ class BoardEncoder(nn.Module):
                 self.num_legal_moves_projection(num_legal_moves.unsqueeze(-1).to(torch.float32)),
                 self.white_remaining_time_projection(white_remaining_time.unsqueeze(-1).to(torch.float32)),
                 self.black_remaining_time_projection(black_remaining_time.unsqueeze(-1).to(torch.float32)),
-                self.time_control_projection(time_control).unsqueeze(1).to(torch.float32),
+                self.time_control_projection(time_control.to(torch.float32)).unsqueeze(1).to(torch.float32),
                 self.phase_embeddings(phase),
                 #self.white_rating_embeddings(white_rating.unsqueeze(-1).to(torch.float32)),
                 #self.black_rating_embeddings(black_rating.unsqueeze(-1).to(torch.float32)),
